tester-service/requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
kafka-python>=2.0.2
prometheus-client>=0.19.0
jinja2>=3.1.2
colorama>=0.4.6
asyncio>=3.4.3
python-dateutil>=2.8.2
